refactor : 수강신청 목록 클래스 생성 후 포장
